+++ Runtime/Interaction/IStatReceiver.cs
@@ -1,0 +1,10 @@
+namespace Dave6.StatSystem.Interaction
+{
+    /// <summary>
+    /// Character, NPC
+    /// </summary>
+    public interface IStatReceiver
+    {
+        void Accept(IStatInvoker invoker);
+    }
+}