+++ Runtime/ProtoCode/StatController.cs
@@ -1,11 +1,14 @@
 using System.Collections.Generic;
 using Dave6.StatSystem;
+using Dave6.StatSystem.Effect;
+using Dave6.StatSystem.Interaction;
+using Dave6.StatSystem.Stat;
 using Unity.Collections;
 using UnityEngine;
 
 namespace ProtoCode
 {
-    public class StatController : MonoBehaviour, IEntity
+    public class StatController : MonoBehaviour, IEntity, IStatReceiver
     {
         [SerializeField] StatDatabase m_StatDatabase;
         public StatDatabase statDatabase => m_StatDatabase;
@@ -13,7 +16,10 @@ namespace ProtoCode
         StatHandler m_StatHandler;
         public StatHandler statHandler => m_StatHandler;
 
+        public ResourceStat health { get; set; }
 
+
+        // 디버깅 이후에 사리질 수 있음
         [SerializeField] EffectDefinition testEffect;
 
 
@@ -65,13 +71,18 @@ namespace ProtoCode
             //     Debug.Log($"After Effect -> Health: {healthStat.currentValue}/{healthStat.finalValue}");
             // }
 
-            {
-                var healthStat = m_StatHandler.GetStat("R_Health") as ResourceStat;
-                Debug.Log($"Before Effect -> Health: {healthStat.currentValue}/{healthStat.finalValue}");
-                m_StatHandler.ApplyEffect(testEffect, healthStat);
-                Debug.Log($"After Effect -> Health: {healthStat.currentValue}/{healthStat.finalValue}");
-            }
+            // {
+            //     var healthStat = m_StatHandler.GetStat("R_Health") as ResourceStat;
+            //     Debug.Log($"Before Effect -> Health: {healthStat.currentValue}/{healthStat.finalValue}");
+            //     m_StatHandler.ApplyEffect(testEffect, healthStat);
+            //     Debug.Log($"After Effect -> Health: {healthStat.currentValue}/{healthStat.finalValue}");
+            // }
 
+        }
+
+        public void Accept(IStatInvoker invoker)
+        {
+            invoker.Invoke(this);
         }
     }
 }